minor tweaks to Poi shaders
vertex color masks have been removed, and the "RenamePoiTagsPropertyFields" script has been added. It interacts with Poi shader tags in Hue/Saturation change anims (Poi adds a tag when you press Lock in Optimized Shader). This is done to avoid the hassle of overwriting animations when new tags are possible.