bind twist bones to the forearm, not to the hand
in the game itself, it’s not the hand that rotates, but the forearm (because of IK). At the moment, unfortunately, the twist bones aren’t working as they should. I might have to abandon this idea if I don’t find a solution and restore the project backup from before this change.